+++ changelog.md
@@ -15,7 +15,7 @@
 	- A significant increase in search speed, especially when dealing with large amounts of data.
 	- Faster indexing.
 	- Word stemming configurable for all languages.
-- Added MetaProperty to control number input grouping 
+- Added `data-invariant` attribute to number input controls 
 
 ### Bugfixes